requirements.txt
# CWRU Time-Frequency Analysis
# Python Dependencies

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Wavelet transforms
PyWavelets>=1.3.0

# Image processing and visualization
Pillow>=9.0.0
matplotlib>=3.5.0

# Type hints support (for Python < 3.9)
typing-extensions>=4.0.0; python_version < "3.9"

# Optional: For faster processing
# numba>=0.56.0
